+++ src/amplitude-snippet.js
@@ -9,11 +9,11 @@
   var as = document.createElement('script');
   as.type = 'text/javascript';
   // Don't edit as.integrity, it is tracked by semantic-release-bot during releases
-  as.integrity = 'sha384-UcvEbHmT0LE2ZB30Y3FmY3Nfw6puAKXz/LpCFuoywywYikMOr/519Uu1yNq2nL9w';
+  as.integrity = 'sha384-uviuD0ZRJOX2rH1yotgp/E1VMF2DeDo/ycJSm6scGkb0jnz5tqsJOXZK6paV8Y14';
   as.crossOrigin = 'anonymous';
   as.async = true;
   // Don't edit as.src, it is tracked by semantic-release-bot during releases
-  as.src = 'https://cdn.amplitude.com/libs/amplitude-8.12.0-min.gz.js';
+  as.src = 'https://cdn.amplitude.com/libs/amplitude-8.13.1-min.gz.js';
   as.onload = function () {
     if (!window.amplitude.runQueuedFunctions) {
       console.log('[Amplitude] Error: could not load SDK');
@@ -75,6 +75,7 @@
     'regenerateDeviceId',
     'groupIdentify',
     'onInit',
+    'onNewSessionStart',
     'logEventWithTimestamp',
     'logEventWithGroups',
     'setSessionId',